pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "m365seed"
version = "0.1.0"
description = "Safe, idempotent seeding tool for Microsoft 365 demo tenants with synthetic, theme-aware HLS content"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "HLS Demo Team"},
]

dependencies = [
    "azure-identity>=1.15.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "typer[all]>=0.9.0",
    "jinja2>=3.1.0",
    "jsonschema>=4.20.0",
    "rich>=13.0.0",
    "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.21.0",
    "coverage>=7.0",
]

[project.scripts]
m365seed = "m365seed.cli:app"

[tool.setuptools.packages.find]
include = ["m365seed*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
